Fixed Splash/Intro names, formatting, small design fixes
