+++ src/views/Select.vue
@@ -1,0 +1,15 @@
+<template>
+    <div id="select">
+
+    </div>
+</template>
+
+<script>
+    export default {
+        name: 'select'
+    }
+</script>
+
+<style lang="scss" scoped>
+
+</style>
